function-calling-json-schema/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
jsonschema>=4.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
openai>=1.0.0 